SB-711 Finding artifacts by it's coordinates (generic implementation).

+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/services/ArtifactEntryServiceTest.java
@@ -14,7 +14,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -27,9 +26,9 @@ import static org.junit.Assert.*;
  * @author Alex Oreshkevich
  * @see https://dev.carlspring.org/youtrack/issue/SB-711
  */
-@Transactional
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = { StorageApiConfig.class })
+@Transactional
 @Rollback(false)
 public class ArtifactEntryServiceTest
 {
@@ -39,23 +38,52 @@ public class ArtifactEntryServiceTest
     @Inject
     ArtifactEntryService artifactEntryService;
 
-    @Autowired
+    @Inject
     OObjectDatabaseTx databaseTx;
 
+    final String storageId = "storage0";
+    final String repositoryId = "release";
+
+    final String groupId = "org.carlspring.strongbox";
+    final String artifactId = "coordinates-test";
+
     @Before
-    public void prepareTests()
+    public synchronized void init()
+            throws Exception
     {
 
-        // uncomment that if you want to debug db entries manually and set @Rollback(false)
-        artifactEntryService.deleteAll();
+        logger.info("[prepareTests] Create artifacts....");
+
+        createArtifacts(groupId, artifactId, storageId, repositoryId);
+        displayAllEntries();
     }
 
     @Test
-    public void testThatArtifactEntryIsCreatable()
+    public synchronized void testAll()
+            throws Exception
     {
+        logger.info("[testAll] Start testing....");
 
-        final String storageId = "storage0";
-        final String repositoryId = "release";
+        if (artifactEntryService.count() > 0)
+        {
+
+            checkThatArtifactEntryIsCreatable();
+
+            searchBySingleCoordinate(groupId, 2);
+            searchByTwoCoordinate(groupId, artifactId, 1);
+
+            artifactEntryService.deleteAll();
+        }
+        else
+        {
+            logger.warn("Artifact entries storage was not initialized properly. Unable to execute any tests.");
+        }
+    }
+
+    private void checkThatArtifactEntryIsCreatable()
+    {
+        final String storageId = "storage3";
+        final String repositoryId = "release432";
 
         ArtifactEntry artifactEntry = createArtifactEntry(createMavenArtifactCoordinates(), storageId, repositoryId);
         logger.info("Saved entity " + artifactEntry);
@@ -64,14 +92,17 @@ public class ArtifactEntryServiceTest
         {
             ArtifactEntry savedEntry = databaseTx.detachAll(
                     artifactEntryService
-                            .findAll()
-                            .orElseThrow(() -> new NullPointerException("Unable to find any artifact entry"))
-                            .get(0), true);
+                            .findOne(artifactEntry.getObjectId())
+                            .orElseThrow(() -> new NullPointerException("Unable to find any artifact entry")), true);
 
-            logger.info("Detached entity " + savedEntry);
+            logger.info("[checkThatArtifactEntryIsCreatable] Detached entity " + savedEntry);
 
             assertEquals(storageId, savedEntry.getStorageId());
             assertEquals(repositoryId, savedEntry.getRepositoryId());
+
+            String savedObjectId = savedEntry.getObjectId();
+            logger.info("[checkThatArtifactEntryIsCreatable] Delete entity by ID " + savedObjectId);
+            artifactEntryService.delete(savedObjectId);
         }
         else
         {
@@ -79,39 +110,26 @@ public class ArtifactEntryServiceTest
         }
     }
 
-
-    /*
-        there will be multiple implementations of ArtifactEntry - MavenArtifactEntry, NugetArtifactEntry, etc...
-        each of these will have it's own *ArtifactCooridnates -- MavenArtifactCoordinates, etc...
-
-        meaning that they will have different cooridnates.
-        as in "fields" for these coordinates.
-        in Maven these will be GAVTC (groupId, artifactId, version, type, classifier)
-        in Nuget these will be id, version, extension, etc...
-        npm will have it's own coordinate fields and so on...
-
-        the idea is to have a generic enough implementation that suits this.
+    /**
+     * Make sure that we are able to search artifacts by single coordinate.
+     *
+     * @param groupId
+     * @throws Exception
      */
-    @Test
-    public void testSearchByCoordinates()
+    private void searchBySingleCoordinate(String groupId,
+                                          final int expectedResultCount)
             throws Exception
     {
 
-        // prepare ArtifactEntry instance in database
-        final String storageId = "storage0";
-        final String repositoryId = "release";
-        createArtifactEntry(createMavenArtifactCoordinates(), storageId, repositoryId);
-
-        // display current database entries
-        displayAllEntries();
-
         // prepare search query key (coordinates)
         MavenArtifactCoordinates query = new MavenArtifactCoordinates();
-        query.setGroupId("org.carlspring.strongbox");
+        query.setGroupId(groupId);
 
         List<ArtifactEntry> result = artifactEntryService.findByCoordinates(query);
         assertNotNull(result);
         assertFalse(result.isEmpty());
+
+        assertEquals(expectedResultCount, result.size());
 
         result.forEach(artifactEntry ->
                        {
@@ -120,11 +138,53 @@ public class ArtifactEntryServiceTest
                            ArtifactEntry artifact = databaseTx.detachAll(artifactEntry, true);
 
                            logger.info("Found artifact " + artifact);
+
+                           assertEquals(groupId, artifact.getArtifactCoordinates()
+                                                         .getCoordinate("groupId"));
+                       });
+    }
+
+    /**
+     * Make sure that we are able to search artifacts by two coordinates that need to be joined with logical AND operator.
+     *
+     * @param groupId
+     * @param artifactId
+     */
+    private void searchByTwoCoordinate(String groupId,
+                                       String artifactId,
+                                       final int expectedResultCount)
+            throws Exception
+    {
+
+        // prepare search query key (coordinates)
+        MavenArtifactCoordinates query = new MavenArtifactCoordinates();
+        query.setGroupId(groupId);
+        query.setArtifactId(artifactId);
+
+        List<ArtifactEntry> result = artifactEntryService.findByCoordinates(query);
+        assertNotNull(result);
+        assertFalse(result.isEmpty());
+
+        assertEquals(expectedResultCount, result.size());
+
+        result.forEach(artifactEntry ->
+                       {
+
+                           databaseTx.activateOnCurrentThread();
+                           ArtifactEntry artifact = databaseTx.detachAll(artifactEntry, true);
+
+                           logger.info("Found artifact " + artifact);
+
+                           assertEquals(groupId, artifact.getArtifactCoordinates()
+                                                         .getCoordinate("groupId"));
+                           assertEquals(artifactId, artifact.getArtifactCoordinates()
+                                                            .getCoordinate("artifactId"));
                        });
     }
 
     private void displayAllEntries()
     {
+        logger.info("[displayAllEntries] ->>>> ...... ");
         List<ArtifactEntry> result = artifactEntryService.findAll()
                                                          .orElse(null);
         if (result == null || result.isEmpty())
@@ -142,6 +202,36 @@ public class ArtifactEntryServiceTest
                        });
     }
 
+    private void createArtifacts(String groupId,
+                                 String artifactId,
+                                 String storageId,
+                                 String repositoryId)
+    {
+        // create 3 artifacts, one will have coordinates that matches our query, one - not
+
+        ArtifactCoordinates coordinates1 = new MavenArtifactCoordinates(groupId,
+                                                                        artifactId + "123",
+                                                                        "1.2.3",
+                                                                        null,
+                                                                        "jar");
+
+        ArtifactCoordinates coordinates2 = new MavenArtifactCoordinates(groupId,
+                                                                        artifactId,
+                                                                        "1.2.3",
+                                                                        null,
+                                                                        "jar");
+
+        ArtifactCoordinates coordinates3 = new MavenArtifactCoordinates(groupId + "myId",
+                                                                        artifactId + "321",
+                                                                        "1.2.3",
+                                                                        null,
+                                                                        "jar");
+
+        createArtifactEntry(coordinates1, storageId, repositoryId);
+        createArtifactEntry(coordinates2, storageId, repositoryId);
+        createArtifactEntry(coordinates3, storageId, repositoryId);
+    }
+
     private ArtifactEntry createArtifactEntry(ArtifactCoordinates coordinates,
                                               String storageId,
                                               String repositoryId)
@@ -149,8 +239,8 @@ public class ArtifactEntryServiceTest
 
         ArtifactEntry artifactEntry = new ArtifactEntry();
         artifactEntry.setArtifactCoordinates(coordinates);
-        artifactEntry.setStorageId("storage0");
-        artifactEntry.setRepositoryId("release");
+        artifactEntry.setStorageId(storageId);
+        artifactEntry.setRepositoryId(repositoryId);
 
         databaseTx.activateOnCurrentThread();
         return databaseTx.detachAll(artifactEntryService.save(artifactEntry), true);
@@ -159,8 +249,8 @@ public class ArtifactEntryServiceTest
     private ArtifactCoordinates createMavenArtifactCoordinates()
     {
 
-        return new MavenArtifactCoordinates("org.carlspring.strongbox",
-                                            "coordinates-test",
+        return new MavenArtifactCoordinates("org.carlspring.strongbox.another.package",
+                                            "coordinates-test-super-test",
                                             "1.2.3",
                                             null,
                                             "jar");
